Workshop_02/requirements.txt
pandas
openai
chardet
python-dotenv
tenacity
streamlit
streamlit-chat
